# docs/21_openapi.yaml
openapi: 3.0.3
info:
  title: Keiba AI API
  version: 0.2.0
  description: |
    競馬AI（keiba.go.jp 由来データのスクレイピング → DB 蓄積 → 参照）のAPI。

    - 本リリースのスコープ: reverse-proxy / api-service / scraper-service / postgres
    - スコープ外: pipeline-service(Prefect), worker-service(Celery), MLflow

    ## Authentication (JWT)
    - Access Token: JWT（短命 15分）を `Authorization: Bearer <token>` で送信
    - Refresh Token: JWT（長命 30日）を HttpOnly Cookie（`refresh_token`）に保存
      - Rotation: refresh のたびに refresh token を更新（毎回更新）
      - Revocation: Redis で失効管理（jti ベース想定）

    ## Scraper authentication (reserved)
    - api-service ⇔ scraper-service 間は **認証不要**（将来のアップデートで対応予定）
    - client_credentials フローは将来予約（後方互換で有効化）
    - 将来的に **mTLS を導入予定**（現時点は未対応）
servers:
  - url: /api
...
paths:
...
  /scrape/manual-tasks:
    post:
      tags: [scrape]
      summary: Request manual scrape task
      description: |
        frontend などから手動スクレイプを依頼する。
        `race_date` を省略した場合は当日（JST）を対象とする。
        api-service は Pi 側へ手動実行を指示する。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualScrapeTaskRequest"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualScrapeTaskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /scrape/schedule:
    get:
      tags: [scrape]
      summary: Get scrape schedule status
      description: |
        定期実行のオン/オフ状態と対象競馬場を返す。
        cron コンテナは本状態に従って実行する。
        認証不要（PC 経由の内部通信を想定）
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeScheduleStatus"
    post:
      tags: [scrape]
      summary: Update scrape schedule
      description: |
        定期実行のオン/オフと対象競馬場を更新する。
        api-service が scraper-service の control API へ反映する。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeScheduleUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeScheduleStatus"
  /scrape/sync:
    post:
      tags: [scrape]
      summary: Trigger manual sync
      description: |
        差分同期（fingerprint ベース）を手動で開始する。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeSyncRequest"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeSyncResponse"
  /scrape/sync/scheduled:
    post:
      tags: [scrape]
      summary: Trigger scheduled sync
      description: |
        定期同期（JST 1日おき）を cron からトリガする。
        スケジュールが無効または未到来の場合はスキップする。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeSyncRequest"
      responses:
        "200":
          description: Skipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeSyncResponse"
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeSyncResponse"
  /scrape/sync/schedule:
    post:
      tags: [scrape]
      summary: Update sync schedule
      description: |
        定期同期の有効/無効と interval_days を更新する。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeSyncScheduleRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeSyncStatus"
...
  /scrape/races:
    post:
      tags: [scrape]
      summary: Upsert races (RaceList)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaceUpsertBatchRequest"
      responses:
        "201":
          description: Created/Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUpsertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /scrape/race-entries:
    post:
      tags: [scrape]
      summary: Upsert race entries (DebaTable)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaceEntryUpsertBatchRequest"
      responses:
        "201":
          description: Created/Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUpsertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /scrape/odds-snapshots:
    post:
      tags: [scrape]
      summary: Upsert odds snapshots and items
      description: |
        オッズは発走時刻基準の `snapshot_kind`（t_minus_60m / t_minus_30m / t_minus_20m / t_minus_10m / t_minus_5m / t_minus_1m / final）でスナップショット保存する。
        手動取得は `captured_at` と `races.start_time` から最も近い `snapshot_kind` を割り当てる（`captured_at >= races.start_time` は `final`）。
        同一 `race_key × bet_type × snapshot_kind × odds_flg` は 1件として上書きする。
        `snapshot_kind` は運用で拡張可能（OpenAPI では固定 enum としない）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OddsSnapshotUpsertRequest"
      responses:
        "201":
          description: Created/Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OddsSnapshotUpsertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /scrape/race-results:
    post:
      tags: [scrape]
      summary: Upsert race results (RaceMarkTable)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaceResultUpsertBatchRequest"
      responses:
        "201":
          description: Created/Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUpsertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /scrape/payouts:
    post:
      tags: [scrape]
      summary: Upsert payouts (RefundMoneyList)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayoutUpsertBatchRequest"
      responses:
        "201":
          description: Created/Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUpsertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /scrape/race-changes:
    post:
      tags: [scrape]
      summary: Insert race changes (cancel/jockey change etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaceChangeInsertBatchRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUpsertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /scrape/raw-fetch-logs:
    post:
      tags: [scrape]
      summary: Insert raw fetch logs (optional raw layer)
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RawFetchLogInsertBatchRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUpsertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: bad_request
            message:
              type: string
              example: invalid payload
            details:
              type: object
              additionalProperties: true
...
    RaceKey:
      type: object
      required: [race_date, baba_code, race_no]
      properties:
        race_date:
          type: string
          format: date
        baba_code:
          type: integer
          format: int32
          description: 競馬場コード
          example: 20
        race_no:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
          example: 3
    BetType:
      type: string
      enum:
        - tansho
        - fukusho
        - wakuren
        - wakutan
        - umaren
        - umatan
        - wide
        - sanrenpuku
        - sanrentan
    SnapshotKind:
      type: string
      description: |
        代表オッズ時点（デフォルト: t_minus_60m / t_minus_30m / t_minus_20m / t_minus_10m / t_minus_5m / t_minus_1m / final）。
        運用で拡張可能なため、未知の値も許容する。
      example: t_minus_10m

    # -------------------------
    # Scrape control / sync
    # -------------------------
    ManualScrapeTaskRequest:
      type: object
      required: [baba_code]
      properties:
        event_id:
          type: string
          format: uuid
          nullable: true
          description: Idempotency key (optional)
        race_date:
          type: string
          format: date
          description: YYYY-MM-DD (JST). Omit to use today (JST).
          nullable: true
        baba_code:
          type: integer
          format: int32
          description: 競馬場コード
        race_no:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
          nullable: true
        reason:
          type: string
          nullable: true
    ManualScrapeTaskResponse:
      type: object
      required: [task_id, status, accepted_at]
      properties:
        task_id:
          type: string
        status:
          type: string
          example: accepted
        accepted_at:
          type: string
          format: date-time
    ScrapeScheduleStatus:
      type: object
      required: [enabled, updated_at]
      properties:
        enabled:
          type: boolean
        baba_codes:
          type: array
          items:
            type: integer
            format: int32
          description: 定期実行の対象競馬場コード（複数指定）
          nullable: true
        mode:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
        note:
          type: string
          nullable: true
    ScrapeScheduleUpdateRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        baba_codes:
          type: array
          items:
            type: integer
            format: int32
          description: 定期実行の対象競馬場コード（複数指定）
          nullable: true
    ScrapeSyncRequest:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
          nullable: true
          description: Idempotency key (optional)
        reason:
          type: string
          nullable: true
    ScrapeSyncScheduleRequest:
      type: object
      required: [enabled, interval_days]
      properties:
        enabled:
          type: boolean
        interval_days:
          type: integer
          minimum: 1
          example: 2
        diff_enabled:
          type: boolean
          nullable: true
    ScrapeSyncResponse:
      type: object
      required: [sync_id, status, started_at]
      properties:
        sync_id:
          type: string
        status:
          type: string
          example: accepted
        started_at:
          type: string
          format: date-time
    ScrapeSyncStatus:
      type: object
      required: [enabled, interval_days, diff_enabled]
      properties:
        enabled:
          type: boolean
        interval_days:
          type: integer
          minimum: 1
          example: 2
        diff_enabled:
          type: boolean
        last_synced_at:
          type: string
          format: date-time
          nullable: true
        next_scheduled_at:
          type: string
          format: date-time
          nullable: true
        last_fingerprint:
          type: string
          nullable: true
        last_attempted_at:
          type: string
          format: date-time
          nullable: true
        last_status:
          type: string
          nullable: true
        last_error:
          type: string
          nullable: true
        last_trigger:
          type: string
          nullable: true
        schedule_updated_at:
          type: string
          format: date-time
          nullable: true

    # -------------------------
    # Admin: OAuth Clients
    # -------------------------
...
    BatchUpsertResponse:
      type: object
      required: [accepted, upserted]
      properties:
        accepted:
          type: integer
          example: 10
        upserted:
          type: integer
          example: 10
        warnings:
          type: array
          items:
            type: string
          nullable: true
    RaceUpsert:
      type: object
      required: [race_key]
      properties:
        race_key:
          $ref: "#/components/schemas/RaceKey"
        start_time:
          type: string
          format: time
          nullable: true
        distance_m:
          type: integer
          format: int32
          nullable: true
        course:
          type: string
          nullable: true
        weather:
          type: string
          nullable: true
        track_condition:
          type: string
          nullable: true
        race_name:
          type: string
          nullable: true
        field_size:
          type: integer
          format: int32
          nullable: true
        status:
          type: string
          nullable: true
    RaceUpsertBatchRequest:
      type: object
      required: [event_id, items]
      properties:
        event_id:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RaceUpsert"
    RaceEntryUpsert:
      type: object
      required: [race_key, horse_number, horse_name]
      properties:
        race_key:
          $ref: "#/components/schemas/RaceKey"
        horse_id:
          type: integer
          format: int64
          nullable: true
        post_position:
          type: integer
          format: int32
          nullable: true
        horse_number:
          type: integer
          format: int32
        horse_name:
          type: string
        jockey_name:
          type: string
          nullable: true
        trainer_name:
          type: string
          nullable: true
        handicap_kg:
          type: number
          format: float
          nullable: true
        body_weight:
          type: integer
          format: int32
          nullable: true
        body_weight_diff:
          type: integer
          format: int32
          nullable: true
    RaceEntryUpsertBatchRequest:
      type: object
      required: [event_id, items]
      properties:
        event_id:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RaceEntryUpsert"
    OddsItemUpsert:
      type: object
      required: [legs, is_ordered]
      properties:
        legs:
          type: array
          items:
            type: integer
            format: int32
        is_ordered:
          type: boolean
        odds_min:
          type: number
          format: float
          nullable: true
        odds_max:
          type: number
          format: float
          nullable: true
        popularity:
          type: integer
          nullable: true
        raw_text:
          type: string
          nullable: true
    OddsSnapshotUpsertRequest:
      type: object
      required: [event_id, race_key, bet_type, snapshot_kind, captured_at, source_url, items]
      properties:
        event_id:
          type: string
          format: uuid
        race_key:
          $ref: "#/components/schemas/RaceKey"
        bet_type:
          $ref: "#/components/schemas/BetType"
        snapshot_kind:
          $ref: "#/components/schemas/SnapshotKind"
        captured_at:
          type: string
          format: date-time
        source_url:
          type: string
        odds_flg:
          type: integer
          description: 表示モード識別子（NULL 可）。一意キーに含める。
          nullable: true
        is_final:
          type: boolean
          default: false
        items:
          type: array
          items:
            $ref: "#/components/schemas/OddsItemUpsert"
    OddsSnapshotUpsertResponse:
      type: object
      required: [race_id, bet_type, snapshot_kind, odds_snapshot_id, num_items]
      properties:
        race_id:
          type: integer
          format: int64
        bet_type:
          $ref: "#/components/schemas/BetType"
        snapshot_kind:
          $ref: "#/components/schemas/SnapshotKind"
        odds_snapshot_id:
          type: integer
          format: int64
        num_items:
          type: integer
          minimum: 0
    RaceResultUpsert:
      type: object
      required: [race_key, finish_position]
      properties:
        race_key:
          $ref: "#/components/schemas/RaceKey"
        finish_position:
          type: integer
          format: int32
        horse_number:
          type: integer
          format: int32
          nullable: true
        time_str:
          type: string
          nullable: true
        margin:
          type: string
          nullable: true
        last3f:
          type: number
          format: float
          nullable: true
        popularity:
          type: integer
          nullable: true
        corner1:
          type: string
          nullable: true
        corner2:
          type: string
          nullable: true
        corner3:
          type: string
          nullable: true
        corner4:
          type: string
          nullable: true
    RaceResultUpsertBatchRequest:
      type: object
      required: [event_id, items]
      properties:
        event_id:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RaceResultUpsert"
    PayoutUpsert:
      type: object
      required: [race_key, bet_type, legs, is_ordered]
      properties:
        race_key:
          $ref: "#/components/schemas/RaceKey"
        bet_type:
          $ref: "#/components/schemas/BetType"
        legs:
          type: array
          items:
            type: integer
            format: int32
        is_ordered:
          type: boolean
        payout_yen:
          type: integer
          format: int32
          nullable: true
        popularity:
          type: integer
          nullable: true
    PayoutUpsertBatchRequest:
      type: object
      required: [event_id, items]
      properties:
        event_id:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PayoutUpsert"
    RaceChangeInsert:
      type: object
      required: [race_key, change_type, payload, captured_at]
      properties:
        race_key:
          $ref: "#/components/schemas/RaceKey"
        change_type:
          type: string
        payload:
          type: object
          additionalProperties: true
        captured_at:
          type: string
          format: date-time
    RaceChangeInsertBatchRequest:
      type: object
      required: [event_id, items]
      properties:
        event_id:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RaceChangeInsert"
    RawFetchLogInsert:
      type: object
      required: [page_type, url, http_status, captured_at]
      properties:
        race_key:
          $ref: "#/components/schemas/RaceKey"
          nullable: true
        page_type:
          type: string
        url:
          type: string
        http_status:
          type: integer
          format: int32
        sha256:
          type: string
          nullable: true
        storage_path:
          type: string
          nullable: true
        captured_at:
          type: string
          format: date-time
        note:
          type: string
          nullable: true
    RawFetchLogInsertBatchRequest:
      type: object
      required: [event_id, items]
      properties:
        event_id:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RawFetchLogInsert"